Add tests to Podcasts (fix bug 844540)
Add some basic UI tests to Podcasts including search and subscription
handling. More tests are needed, but now the groundwork is there.

This also includes CasperJS, a simple node server to serve the app, a
casperjs runner that ignores web security (to make "systemXHR"-style calls),
and an IndexedDB polyfill over WebSQL so the app can behave as it would
on a device. *Phew!*